remove premature optimization of useEffect()
fixes https://github.com/pmuellr/us-weather/issues/4

Change `useEffect()` usage with event handlers to not pass the empty
array of changed properties.  So, it will run on every update now.

Also changed `use-node-event` to `use-on-off-event`, as the hook is
not specific to node, but general to all event handling targets that
use `on()` and `off()` for listeners (like Node, but also Leaflet!).